+++ Changelog.md
@@ -1,4 +1,13 @@
 # Changelog
 
+## 0.2.0
+
+Release 0.2.0 has been developed to be compatible with `transitions` 0.8.1 and contains the following changes:
+
+* update cytoscape to 3.15.0
+* added enter/exit markup and conditions; pass `details=true` as an URL argument to activate
+* added handling of internal transitions
+
 ## 0.1.0
 
+initial release